// src/enum/category-ingredient.enum.ts
/* eslint-disable prettier/prettier */
export enum CategoryIngredient {

    ACEITES_GRASAS = "Aceites y grasas",
    CARNE_PESCADO = "Carne y pescado",
    HORNEAR = "Cereales y frutos secos",
    FRUTAS = "Frutas",
    HIERBAS_ESPECIES = "Hierbas y especies",
    HUEVOS_LACTEOS = "Huevos y lacteos",
    VERDURAS = "Verduras",
    PASTA_ARROZ_LEGUMBRES = "Pasta, arroz y legumbres",
    OTROS = "Otros"

}